requirements.txt
# 核心数据与计算
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=15.0.0      # parquet 读写支持

# 配置解析
pyyaml>=6.0

# 技术指标计算
ta-lib>=0.4.28        # 需额外安装 TA-Lib C 库，参考 docs/

# 机器学习（可选）
scikit-learn>=1.4.0
lightgbm>=4.3.0

# 可视化
matplotlib>=3.8.0

# 测试
pytest>=8.0.0
pytest-cov>=5.0.0
